+++ src/daf_gui/tabs/__init__.py
@@ -2,4 +2,5 @@ from .analytics import *
 from .optional import *
 from .schema import *
 from .about import *
+from .debug import *
 from .live import *
